Add CJK number parser for Python
