+++ Python/libraries/recognizers-number/recognizers_number/number/__init__.py
@@ -5,5 +5,6 @@ from .english import *
 from .spanish import *
 from .chinese import *
 from .french import *
+from .japanese import *
 from .number_recognizer import *
 from .parser_factory import *
